requirements.txt
# StealthRL Requirements - Tinker Integration

# Core Tinker dependencies
tinker  # Tinker API client
tinker-cookbook  # Tinker Cookbook utilities

# PyTorch (for reward computation and model utilities)
torch>=2.0.0
transformers>=4.30.0

# Sentence transformers (for semantic similarity)
sentence-transformers>=2.2.0

# BERTScore (for semantic similarity evaluation)
bert-score>=0.3.13

# Text processing
textstat>=0.7.3  # Readability metrics

# Database (for detector caching)
# sqlite3 is built-in to Python, no need to install

# Utilities
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
tqdm>=4.65.0

# Evaluation
scikit-learn>=1.3.0  # For AUROC, F1, etc.
tiktoken>=0.5.0  # For token-based text chunking

# Visualization
matplotlib>=3.7.0  # For plotting
seaborn>=0.12.0  # For statistical visualizations

# Logging and monitoring
tensorboard>=2.13.0  # Optional: for visualization
wandb>=0.15.0  # Optional: for experiment tracking

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Async utilities
aiofiles>=23.0.0
httpx>=0.24.0  # For async HTTP requests

# Original StealthRL dependencies (from requirements.txt)
# These may be needed for baseline comparisons
# trl>=0.7.0  # Not needed for Tinker (replaced by tinker-cookbook)
# peft>=0.5.0  # Not needed for Tinker (LoRA handled internally)
# datasets>=2.14.0
# accelerate>=0.23.0  # Not needed for Tinker

# Optional: for specific detectors
# fast-detectgpt>=0.1.0  # If using Fast-DetectGPT locally
# ghostbuster>=0.1.0  # If using Ghostbuster locally
# binoculars>=0.1.0  # If using Binoculars locally

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0